Refactored the teacher report
